requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.26
django-cors-headers==4.9.0
django-model-utils==5.0.0
django-prometheus==2.4.1
djangorestframework==3.16.1
idna==3.11
pika==1.3.2
prometheus_client==0.23.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tenacity==9.1.2
urllib3==2.5.0
drf_spectacular
